backend/requirements.txt
fastapi==0.115.3
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
python-multipart==0.0.9
pandas==2.2.3
openpyxl==3.1.5
alembic==1.13.3
pydantic-settings==2.4.0
redis==5.1.1
celery==5.4.0
google-generativeai==0.7.2
